fast_v1: separate into three backends with varying accuracy level
