+++ .gitignore
@@ -2,7 +2,7 @@
 .vs/
 .vscode/
 .cache/
-build/
+build*/
 cmake/open-cpp-coverage.cmake
 cmake-build-*/
 prefix/
