+++ include/operon/interpreter/backend/fast_v2.hpp
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: Copyright 2019-2024 Heal Research
 
-#ifndef OPERON_BACKEND_FAST_V1_HPP
-#define OPERON_BACKEND_FAST_V1_HPP
+#ifndef OPERON_BACKEND_FAST_V2_HPP
+#define OPERON_BACKEND_FAST_V2_HPP
 
-#define OPERON_MATH_FAST_APPROX_PRECISION 0
+#define OPERON_MATH_FAST_APPROX_PRECISION 1
 #include "fast_approx/derivatives.hpp"
 
 #endif
